smart-agriculture-flutter-app-backend-main/src/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pillow
flask
flask-cors
numpy
tqdm
scikit-learn
